# Pipeline: optimize part scan