--- a/dbms/src/Operators/tests/gtest_concat_source.cpp
+++ b/dbms/src/Operators/tests/gtest_concat_source.cpp
@@ -1,0 +1,40 @@
+// Copyright 2023 PingCAP, Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <Flash/Executor/PipelineExecutorStatus.h>
+#include <Operators/ConcatSourceOp.h>
+#include <TestUtils/ColumnGenerator.h>
+#include <gtest/gtest.h>
+
+namespace DB::tests
+{
+class TestConcatSource : public ::testing::Test
+{
+};
+
+TEST_F(TestConcatSource, setBlockSink)
+{
+    Block res;
+    ASSERT_FALSE(res);
+
+    PipelineExecutorStatus exec_status;
+    SetBlockSinkOp set_block_sink{exec_status, "test", res};
+    Block header{ColumnGenerator::instance().generate({0, "Int32", DataDistribution::RANDOM})};
+    set_block_sink.setHeader(header);
+    Block block{ColumnGenerator::instance().generate({2, "Int32", DataDistribution::RANDOM})};
+    set_block_sink.write(std::move(block));
+    ASSERT_TRUE(res);
+}
+
+} // namespace DB::tests