--- a/dbms/src/Operators/tests/gtest_concat_source.cpp
+++ b/dbms/src/Operators/tests/gtest_concat_source.cpp
@@ -16,9 +16,44 @@
 #include <Operators/ConcatSourceOp.h>
 #include <TestUtils/ColumnGenerator.h>
 #include <gtest/gtest.h>
+#include <memory>
+#include <type_traits>
+#include "Flash/Pipeline/Exec/PipelineExecBuilder.h"
+#include "Operators/Operator.h"
 
 namespace DB::tests
 {
+namespace
+{
+class MockSourceOp : public SourceOp
+{
+public:
+    MockSourceOp(
+        PipelineExecutorStatus & exec_status_,
+        const Block & output_)
+        : SourceOp(exec_status_, "mock")
+        , output(output_)
+    {
+        setHeader(output.cloneEmpty());
+    }
+
+    String getName() const override
+    {
+        return "MockSourceOp";
+    }
+
+protected:
+    OperatorStatus readImpl(Block & block) override
+    {
+        std::swap(block, output);
+        return OperatorStatus::HAS_OUTPUT;
+    }
+
+private:
+    Block output;
+};
+}
+
 class TestConcatSource : public ::testing::Test
 {
 };
@@ -33,8 +68,39 @@ TEST_F(TestConcatSource, setBlockSink)
     Block header{ColumnGenerator::instance().generate({0, "Int32", DataDistribution::RANDOM})};
     set_block_sink.setHeader(header);
     Block block{ColumnGenerator::instance().generate({2, "Int32", DataDistribution::RANDOM})};
+    set_block_sink.operatePrefix();
     set_block_sink.write(std::move(block));
     ASSERT_TRUE(res);
+    set_block_sink.operateSuffix();
+}
+
+TEST_F(TestConcatSource, concatSink)
+{
+    size_t block_cnt = 10;
+    std::vector<PipelineExecBuilder> builders;
+    PipelineExecutorStatus exec_status;
+    for (size_t i = 0; i < block_cnt; ++i)
+    {
+        PipelineExecBuilder builder;
+        Block block{ColumnGenerator::instance().generate({2, "Int32", DataDistribution::RANDOM})};
+        builder.setSourceOp(std::make_unique<MockSourceOp>(exec_status, block));
+        builders.push_back(std::move(builder));
+    }
+
+    ConcatSourceOp concat_source{exec_status, "test", builders};
+    size_t actual_block_cnt = 0;
+    concat_source.operatePrefix();
+    while (true)
+    {
+        Block tmp;
+        concat_source.read(tmp);
+        if (tmp)
+            ++actual_block_cnt;
+        else
+            break;
+    }
+    concat_source.operateSuffix();
+    ASSERT_EQ(actual_block_cnt, block_cnt);
 }
 
 } // namespace DB::tests