### What problem does this PR solve?

Issue Number: ref #6518 ref #4474

Problem Summary:

### What is changed and how it works?
- Avoid excessive memory usage due to too many partitions read concurrently
- Under the storage thread pool + pipeline model, data imbalance between multiple partitions will not cause data skew in the computing layer, so here is simply concat the streams of different partitions.

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
```

create table part_test.no_part_table(a int, b int);

create table part_test.part_table(a int, b int)
PARTITION BY RANGE (a) (
    PARTITION p0 VALUES LESS THAN (1),
    PARTITION p1 VALUES LESS THAN (2),
    ...
    PARTITION p59 VALUES LESS THAN (59),
    PARTITION p60 VALUES LESS THAN (MAXVALUE)
);

ALTER TABLE part_test.no_part_table SET TIFLASH REPLICA 1;
ALTER TABLE part_test.part_table SET TIFLASH REPLICA 1;

60 partition, 50w per partition.

set tidb_enable_tiflash_pipeline_model = ON;
select count(1) from no_part_table;
+----------+
| count(1) |
+----------+
| 30000000 |
+----------+
1 row in set (0.08 sec)
select count(1) from no_part_table where b = 0 or b = 5;
+----------+
| count(1) |
+----------+
|   200000 |
+----------+
1 row in set (0.02 sec)
select count(1) from part_table;
+----------+
| count(1) |
+----------+
| 30000000 |
+----------+
1 row in set (0.43 sec)
select count(1) from part_table where b = 0 or b = 5;
+----------+
| count(1) |
+----------+
|   200000 |
+----------+
1 row in set (0.08 sec)

set tidb_enable_tiflash_pipeline_model = OFF;
select count(1) from no_part_table;
+----------+
| count(1) |
+----------+
| 30000000 |
+----------+
1 row in set (0.07 sec)
select count(1) from no_part_table where b = 0 or b = 5;
+----------+
| count(1) |
+----------+
|   200000 |
+----------+
1 row in set (0.01 sec)
select count(1) from part_table;
+----------+
| count(1) |
+----------+
| 30000000 |
+----------+
1 row in set (0.45 sec)
select count(1) from part_table where b = 0 or b = 5;
+----------+
| count(1) |
+----------+
|   200000 |
+----------+
1 row in set (0.09 sec)
```
- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- bugfix or new feature needs a release note -->

```release-note
None
```